requirement.txt
pandas
numpy
seaborn
matplotlib
langchain_community
langchain_google_genai
google-generativeai
gradio